Fix provided number of occurences > max for variable\n- Up to 65535 occurences possible

+++ Output.h
@@ -6,7 +6,7 @@
 /// @brief Output structure to handle string, number of occurences and process to output it to standart stream.
 struct Output
 {
-    uint8_t nb;         ///< Total number of occurences that str shall be outputed.
+    uint16_t nb;         ///< Total number of occurences that str shall be outputed.
     std::string str;    ///< String to be outputed.
 
     /// @brief Default constructor.
